+++ forui/test/src/widgets/alert/alert_icon_test.dart
@@ -9,7 +9,7 @@ void main() {
     test('invalid height', () {
       expect(
         () => FAlertIconStyle(
-          dimension: 0,
+          size: 0,
           color: Colors.white,
         ),
         throwsAssertionError,
@@ -19,7 +19,7 @@ void main() {
     test('valid height', () {
       expect(
         () => FAlertIconStyle(
-          dimension: 1,
+          size: 1,
           color: Colors.white,
         ),
         returnsNormally,